+++ src/com/parsam/controller/product/ProductDeleteAction.java
@@ -1,4 +1,4 @@
-package com.parsam.controller;
+package com.parsam.controller.product;
 
 import com.parsam.comm.Action;
 import com.parsam.comm.Forward;
@@ -10,18 +10,19 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class ProductDetailAction implements Action {
+public class ProductDeleteAction implements Action {
+
     @Override
     public Forward execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        long pid = Long.parseLong(request.getParameter("pid"));
+        long pid=Long.parseLong(request.getParameter("pid"));
+        String pimg=request.getParameter("pimg");
 
-        ProductService service = ProductService.getService();
-        ProductDTO pdto = service.showDetail(pid);
+        request.setAttribute("pid", pid);
+        request.setAttribute("pimg",pimg);
 
-        request.setAttribute("pdto",pdto);
         Forward forward = new Forward();
         forward.setForward(true);
-        forward.setUrl("WEB-INF/main/index.jsp?page=../chh_product/detail.jsp");
+        forward.setUrl("WEB-INF/main/index.jsp?page=../chh_product/delete_confirm.jsp");
         return forward;
     }
 }
