+++ src/com/parsam/controller/product/ProductDetailAction.java
@@ -1,4 +1,4 @@
-package com.parsam.controller;
+package com.parsam.controller.product;
 
 import com.parsam.comm.Action;
 import com.parsam.comm.Forward;
@@ -10,18 +10,18 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class ProductUpdateAction implements Action {
+public class ProductDetailAction implements Action {
     @Override
     public Forward execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        long pid=Long.parseLong(request.getParameter("pid"));
+        long pid = Long.parseLong(request.getParameter("pid"));
 
         ProductService service = ProductService.getService();
-        ProductDTO dto = service.showDetail(pid);
-        request.setAttribute("dto",dto);
+        ProductDTO pdto = service.showDetail(pid);
 
+        request.setAttribute("pdto",pdto);
         Forward forward = new Forward();
         forward.setForward(true);
-        forward.setUrl("WEB-INF/main/index.jsp?page=../chh_product/updateform.jsp");
+        forward.setUrl("WEB-INF/main/index.jsp?page=../chh_product/detail.jsp");
         return forward;
     }
 }
