+++ src/com/parsam/controller/product/ProductListAction.java
@@ -1,31 +1,28 @@
-package com.parsam.controller;
+package com.parsam.controller.product;
 
 import com.parsam.comm.Action;
 import com.parsam.comm.Forward;
+import com.parsam.dto.ProductDTO;
 import com.parsam.service.ProductService;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.List;
 
-public class ProductDeleteResultAction implements Action {
+public class ProductListAction implements Action {
+
     @Override
     public Forward execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        long pid = Long.parseLong(request.getParameter("pid"));
-        String pimg = request.getParameter("pimg");
-
-        String realpath = request.getServletContext().getRealPath("productUpload");
         ProductService service = ProductService.getService();
-
-        if (pimg==null)
-            pimg="";
-
-        service.deleteData(pid, realpath+"/"+pimg);
+        List<ProductDTO> list = service.productListService();
+        request.setAttribute("list", list);
+        System.out.println(list);
 
         Forward forward = new Forward();
         forward.setForward(true);
-        forward.setUrl("WEB-INF/main/index.jsp?page=../chh_product/delete_alert.jsp");
+        forward.setUrl("WEB-INF/main/index.jsp?page=list.jsp");
         return forward;
     }
 }
