+++ src/com/parsam/controller/product/ProductMainListAction.java
@@ -1,4 +1,4 @@
-package com.parsam.controller;
+package com.parsam.controller.product;
 
 import com.parsam.comm.Action;
 import com.parsam.comm.Forward;
@@ -8,13 +8,12 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class ProductlistResultAction implements Action {
+public class ProductMainListAction implements Action {
     @Override
     public Forward execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-
         Forward forward = new Forward();
         forward.setForward(true);
-        forward.setUrl("WEB-INF/main/index.jsp?page=mainListResult.jsp");
-        return null;
+        forward.setUrl("WEB-INF/main/index.jsp?page=mainList.jsp");
+        return forward;
     }
 }
