+++ src/com/parsam/controller/product/ProductUpdateAlertAction.java
@@ -1,4 +1,4 @@
-package com.parsam.controller;
+package com.parsam.controller.product;
 
 import com.parsam.comm.Action;
 import com.parsam.comm.Forward;
@@ -8,12 +8,15 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class ProductWriteAction implements Action {
+public class ProductUpdateAlertAction implements Action {
     @Override
     public Forward execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        long pid = (Long) request.getAttribute("pid");
+
+        request.setAttribute("pid", pid);
         Forward forward = new Forward();
         forward.setForward(true);
-        forward.setUrl("WEB-INF/main/index.jsp?page=../chh_product/writeform.jsp");
+        forward.setUrl("WEB-INF/main/index.jsp?page=../chh_product/update_alert.jsp");
         return forward;
     }
 }
