+++ src/com/parsam/controller/product/ProductlistResultAction.java
@@ -1,0 +1,20 @@
+package com.parsam.controller.product;
+
+import com.parsam.comm.Action;
+import com.parsam.comm.Forward;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+public class ProductlistResultAction implements Action {
+    @Override
+    public Forward execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
+        Forward forward = new Forward();
+        forward.setForward(true);
+        forward.setUrl("WEB-INF/main/index.jsp?page=mainListResult.jsp");
+        return null;
+    }
+}
