+++ web/WEB-INF/prop.properties
@@ -1,4 +1,4 @@
-#미라
+#\uBBF8\uB77C
 /boardlist.do=com.parsam.controller.board.BoardListAction
 /board_detail.do=com.parsam.controller.board.BoardDetailAction
 /board_delete.do=com.parsam.controller.board.BoardDeleteAction
@@ -6,18 +6,22 @@
 /board_result.do=com.parsam.controller.board.BoardWriteResultAction
 /board_update.do=com.parsam.controller.board.BoardUpdateAction
 /board_update_result.do=com.parsam.controller.board.BoardUpdateResultAction
-/product_write.do=com.parsam.controller.ProductWriteAction
-/product_write_result.do=com.parsam.controller.ProductWriteResultAction
-/product_write_alert.do=com.parsam.controller.ProductWriteAlertAction
-/product_detail.do=com.parsam.controller.ProductDetailAction
-/product_update.do=com.parsam.controller.ProductUpdateAction
-/product_update_result.do=com.parsam.controller.ProductUpdateResultAction
-/product_update_alert.do=com.parsam.controller.ProductUpdateAlertAction
-/product_delete.do=com.parsam.controller.ProductDeleteAction
-/product_delete_result.do=com.parsam.controller.ProductDeleteResultAction
+#\uD604\uD6C8
+/product_write.do=com.parsam.controller.product.ProductWriteAction
+/product_write_result.do=com.parsam.controller.product.ProductWriteResultAction
+/product_write_alert.do=com.parsam.controller.product.ProductWriteAlertAction
+/product_detail.do=com.parsam.controller.product.ProductDetailAction
+/product_update.do=com.parsam.controller.product.ProductUpdateAction
+/product_update_result.do=com.parsam.controller.product.ProductUpdateResultAction
+/product_update_alert.do=com.parsam.controller.product.ProductUpdateAlertAction
+/product_delete.do=com.parsam.controller.product.ProductDeleteAction
+/product_delete_result.do=com.parsam.controller.product.ProductDeleteResultAction
+/product_order.do=com.parsam.controller.product.ProductOrderAction
+#\uCC44\uC6D0
 /index.do=com.parsam.controller.IndexAction
-/list.do=com.parsam.controller.ProductListAction
-/mainList.do=com.parsam.controller.ProductMainListAction
-/mainListResult.do=com.parsam.controller.ProductlistResultAction
+/list.do=com.parsam.controller.product.ProductListAction
+/mainList.do=com.parsam.controller.product.ProductMainListAction
+/mainListResult.do=com.parsam.controller.product.ProductlistResultAction
+#\uC218\uD5A5
 /join.do=com.parsam.controller.user.WriteAction
 /join_result.do=com.parsam.controller.user.WriteResultAction
